# openapi-spec.yaml
openapi: 3.0.3
info:
  title: API Sistema de Calendario
  description: |
    API RESTful para gestión de calendarios con las siguientes características:
    
    ## Arquitectura
    - **Capa de Presentación**: Aplicaciones web, móviles, tablets y servicios externos
    - **Capa de Integración**: API Gateway HTTPS
    - **Capa de Servicios**: Autenticación, Calendario, Notificaciones, Cifrado
    - **Capa de Infraestructura**: Bases de datos, broker de mensajes, proveedores externos
    
    ## Seguridad
    - Autenticación mediante JWT (JSON Web Tokens)
    - Cifrado AES de datos sensibles en reposo
    - Transporte seguro HTTPS
    - Control de acceso basado en permisos
    
    ## Funcionalidades Principales
    1. **Autenticación**: Login con username/password, emisión de token JWT
    2. **Calendarios**: Crear, visualizar calendarios cifrados (requiere autenticación)
    3. **Eventos**: Agregar eventos a calendarios con información cifrada
    4. **Alertas**: Configurar notificaciones (PUSH, EMAIL, SMS) para eventos
    5. **Notificaciones**: Sistema asíncrono con confirmación de entrega

  version: 1.0.0
  contact:
    name: Sistema de Calendario
    email: soporte@calendario.com

servers:
  - url: http://localhost:8080
    description: Servidor de Desarrollo
  - url: https://api.calendario.com
    description: Servidor de Producción (HTTPS)

tags:
  - name: Autenticación
    description: Endpoints para autenticación y registro de usuarios
  - name: Calendario
    description: Endpoints para gestión de calendarios y eventos
  - name: Alertas y Notificaciones
    description: Endpoints para configuración de alertas y notificaciones

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      description: |
        Autentica un usuario mediante username y password.
        Retorna un token JWT válido por 24 horas.
        
        **Flujo**:
        1. Cliente envía credenciales (username + password)
        2. ServicioAutenticacion valida contra BaseDatosUsuarios
        3. Si válido, genera token JWT
        4. Retorna token + información básica del usuario
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              usuario1:
                summary: Usuario de ejemplo
                value:
                  username: "juan.perez"
                  password: "Password123!"
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Login exitoso
                  value:
                    token: "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJqdWFuLnBlcmV6IiwiaWF0IjoxNzAxMDM..."
                    tipo: "Bearer"
                    usuarioId: 1
                    username: "juan.perez"
                    email: "juan.perez@email.com"
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/register:
    post:
      tags:
        - Autenticación
      summary: Registrar nuevo usuario
      description: Registra un nuevo usuario en el sistema y retorna un token JWT
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroRequest'
      responses:
        '200':
          description: Registro exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Error en el registro (username o email ya existe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/token:
    post:
      tags:
        - Autenticación
      summary: Validar token
      description: Valida un token JWT existente
      operationId: validateToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token válido
          content:
            text/plain:
              schema:
                type: string
                example: "Token válido"
        '401':
          description: Token inválido o expirado

  /calendar:
    get:
      tags:
        - Calendario
      summary: Obtener calendarios del usuario
      description: |
        Retorna todos los calendarios del usuario autenticado.
        
        **Flujo**:
        1. Verifica autenticación (token JWT)
        2. ServicioCalendario obtiene calendarios de BD_Calendarios
        3. ServicioCifrado descifra datos sensibles
        4. Retorna vista preparada con datos descifrados
      operationId: getCalendars
      responses:
        '200':
          description: Lista de calendarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarioResponse'
        '401':
          description: No autenticado

    post:
      tags:
        - Calendario
      summary: Crear calendario
      description: |
        Crea un nuevo calendario para el usuario autenticado.
        La descripción se cifra antes de almacenar.
      operationId: createCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarioRequest'
      responses:
        '200':
          description: Calendario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarioResponse'
        '400':
          description: Error en los datos del calendario

  /calendar/view:
    get:
      tags:
        - Calendario
      summary: Mostrar calendario (vista preparada)
      description: |
        Retorna un calendario específico con todos sus eventos.
        Todos los datos cifrados son descifrados antes de enviar la respuesta.
        
        **Control de Acceso**:
        - El usuario debe ser propietario del calendario, O
        - El calendario debe ser público
      operationId: viewCalendar
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID del calendario a visualizar
      responses:
        '200':
          description: Vista del calendario con eventos descifrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarioResponse'
        '403':
          description: Sin permisos para ver este calendario

  /calendar/{id}:
    get:
      tags:
        - Calendario
      summary: Obtener calendario por ID
      description: Retorna un calendario específico si el usuario tiene permisos
      operationId: getCalendarById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Calendario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarioResponse'
        '403':
          description: Sin permisos
        '404':
          description: Calendario no encontrado

  /calendar/events:
    post:
      tags:
        - Calendario
      summary: Agregar evento al calendario
      description: |
        Agrega un nuevo evento a un calendario.
        La descripción del evento se cifra antes de almacenar.
        
        **Validaciones**:
        - El usuario debe ser propietario del calendario
        - fechaFin debe ser posterior a fechaInicio
      operationId: addEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventoRequest'
            examples:
              reunion:
                summary: Reunión de trabajo
                value:
                  titulo: "Reunión con cliente importante"
                  descripcion: "Discutir contrato confidencial XYZ"
                  fechaInicio: "2024-12-01T10:00:00"
                  fechaFin: "2024-12-01T11:30:00"
                  ubicacion: "Sala de conferencias"
                  todoElDia: false
                  tipo: "EVENTO_LABORAL"
                  calendarioId: 1
      responses:
        '200':
          description: Evento creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoResponse'
        '400':
          description: Error en los datos del evento
        '403':
          description: Sin permisos para modificar este calendario

  /alerts:
    get:
      tags:
        - Alertas y Notificaciones
      summary: Obtener alertas del usuario
      description: Retorna todas las alertas configuradas por el usuario autenticado
      operationId: getAlerts
      responses:
        '200':
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertaResponse'

    post:
      tags:
        - Alertas y Notificaciones
      summary: Configurar alerta
      description: |
        Configura una alerta para un evento del calendario.
        
        **Flujo**:
        1. Valida que el usuario tenga acceso al evento
        2. Guarda la alerta en la base de datos
        3. Publica evento en BrokerMensajes (RabbitMQ)
        4. Scheduler procesa alertas pendientes cada minuto
        5. Cuando llega la hora, envía notificación al proveedor correspondiente
        
        **Tipos de Notificación**:
        - PUSH: Notificación push a app móvil
        - EMAIL: Correo electrónico
        - SMS: Mensaje de texto
      operationId: createAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertaRequest'
            examples:
              alertaPush:
                summary: Alerta con notificación Push
                value:
                  eventoId: 5
                  fechaAlerta: "2024-12-01T09:45:00"
                  tipoNotificacion: "PUSH"
                  minutosAnticipacion: 15
                  mensaje: "Recordatorio: Reunión con cliente en 15 minutos"
      responses:
        '200':
          description: Alerta configurada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertaResponse'
        '400':
          description: Error en los datos de la alerta
        '403':
          description: Sin permisos para configurar alerta en este evento

  /notifications/shown:
    post:
      tags:
        - Alertas y Notificaciones
      summary: Confirmar notificación mostrada
      description: |
        Endpoint para que las aplicaciones cliente confirmen que una notificación
        fue mostrada al usuario (confirmación de entrega).
        
        Este endpoint es llamado por:
        - Aplicación Web
        - Aplicación Móvil
        - Aplicación Tablet
        
        La confirmación se registra en el BrokerMensajes para auditoría.
      operationId: notificationShown
      parameters:
        - name: alertaId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la alerta que fue mostrada
      responses:
        '200':
          description: Confirmación recibida
          content:
            text/plain:
              schema:
                type: string
                example: "Confirmación de notificación recibida"
        '400':
          description: Error en la confirmación

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación mediante token JWT.
        
        **Cómo obtener el token**:
        1. Hacer POST a /auth/login con username y password
        2. El servidor retorna un token JWT válido por 24 horas
        3. Incluir el token en todas las peticiones protegidas
        
        **Formato del header**:
        ```
        Authorization: Bearer {token}
        ```

  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "juan.perez"
          description: Nombre de usuario único
        password:
          type: string
          format: password
          minLength: 6
          example: "Password123!"
          description: Contraseña del usuario

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Token JWT para autenticación
          example: "eyJhbGciOiJIUzUxMiJ9..."
        tipo:
          type: string
          default: "Bearer"
          example: "Bearer"
        usuarioId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: "juan.perez"
        email:
          type: string
          format: email
          example: "juan.perez@email.com"

    RegistroRequest:
      type: object
      required:
        - username
        - password
        - email
        - nombre
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "maria.garcia"
        password:
          type: string
          format: password
          minLength: 6
          example: "SecurePass456!"
        email:
          type: string
          format: email
          example: "maria.garcia@email.com"
        nombre:
          type: string
          example: "María García"

    CalendarioRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          example: "Calendario Personal"
          description: Nombre del calendario
        descripcion:
          type: string
          example: "Mis eventos personales y familiares"
          description: Descripción (será cifrada)
        color:
          type: string
          default: "#3788d8"
          example: "#FF5733"
          description: Color en formato hexadecimal
        publico:
          type: boolean
          default: false
          description: Si es público, otros usuarios pueden verlo

    CalendarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: "Calendario Personal"
        descripcion:
          type: string
          example: "Mis eventos personales y familiares"
          description: Descripción descifrada
        color:
          type: string
          example: "#3788d8"
        publico:
          type: boolean
          example: false
        fechaCreacion:
          type: string
          format: date-time
          example: "2024-11-26T10:00:00"
        fechaModificacion:
          type: string
          format: date-time
          example: "2024-11-26T15:30:00"
        eventos:
          type: array
          items:
            $ref: '#/components/schemas/EventoResponse'
        propietarioId:
          type: integer
          format: int64
          example: 1
        propietarioNombre:
          type: string
          example: "Juan Pérez"

    EventoRequest:
      type: object
      required:
        - titulo
        - fechaInicio
        - fechaFin
        - calendarioId
      properties:
        titulo:
          type: string
          example: "Reunión importante"
        descripcion:
          type: string
          example: "Discutir proyecto confidencial"
          description: Descripción (será cifrada)
        fechaInicio:
          type: string
          format: date-time
          example: "2024-12-01T10:00:00"
        fechaFin:
          type: string
          format: date-time
          example: "2024-12-01T11:30:00"
        ubicacion:
          type: string
          example: "Sala de juntas"
        todoElDia:
          type: boolean
          default: false
        tipo:
          type: string
          enum: [REUNION, TAREA, RECORDATORIO, EVENTO_PERSONAL, EVENTO_LABORAL]
          default: "EVENTO_PERSONAL"
        calendarioId:
          type: integer
          format: int64
          example: 1

    EventoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        titulo:
          type: string
          example: "Reunión importante"
        descripcion:
          type: string
          example: "Discutir proyecto confidencial"
          description: Descripción descifrada
        fechaInicio:
          type: string
          format: date-time
          example: "2024-12-01T10:00:00"
        fechaFin:
          type: string
          format: date-time
          example: "2024-12-01T11:30:00"
        ubicacion:
          type: string
          example: "Sala de juntas"
        todoElDia:
          type: boolean
          example: false
        tipo:
          type: string
          example: "EVENTO_LABORAL"
        fechaCreacion:
          type: string
          format: date-time
          example: "2024-11-26T09:00:00"

    AlertaRequest:
      type: object
      required:
        - eventoId
        - fechaAlerta
        - tipoNotificacion
      properties:
        eventoId:
          type: integer
          format: int64
          example: 5
          description: ID del evento para el cual configurar la alerta
        fechaAlerta:
          type: string
          format: date-time
          example: "2024-12-01T09:45:00"
          description: Fecha y hora en que se debe enviar la alerta
        tipoNotificacion:
          type: string
          enum: [PUSH, EMAIL, SMS]
          example: "PUSH"
          description: Tipo de notificación a enviar
        minutosAnticipacion:
          type: integer
          default: 15
          example: 15
          description: Minutos de anticipación al evento
        mensaje:
          type: string
          example: "Recordatorio: Reunión en 15 minutos"
          description: Mensaje personalizado de la alerta

    AlertaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        eventoId:
          type: integer
          format: int64
          example: 5
        eventoTitulo:
          type: string
          example: "Reunión importante"
        fechaAlerta:
          type: string
          format: date-time
          example: "2024-12-01T09:45:00"
        tipoNotificacion:
          type: string
          enum: [PUSH, EMAIL, SMS]
          example: "PUSH"
        enviada:
          type: boolean
          example: false
          description: Indica si la notificación ya fue enviada
        fechaEnvio:
          type: string
          format: date-time
          nullable: true
          example: null
          description: Fecha en que se envió la notificación (null si no se ha enviado)
        minutosAnticipacion:
          type: integer
          example: 15
        mensaje:
          type: string
          example: "Recordatorio: Reunión en 15 minutos"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Error en la solicitud"
        mensaje:
          type: string
          example: "Credenciales inválidas"
        timestamp:
          type: string
          format: date-time
          example: "2024-11-26T10:00:00"